requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt --emit-index-url --universal
--index-url https://pypi.org/simple

asgiref==3.8.1
    # via django
dj-database-url==2.3.0
    # via -r requirements.in
django==5.1.4
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-debug-toolbar
django-debug-toolbar==4.4.6
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
markdown==3.7
    # via -r requirements.in
packaging==24.2
    # via gunicorn
psycopg==3.2.3
    # via -r requirements.in
psycopg-binary==3.2.3 ; implementation_name != 'pypy'
    # via psycopg
psycopg-pool==3.2.4
    # via psycopg
python-dotenv==1.0.1
    # via -r requirements.in
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.12.2
    # via
    #   dj-database-url
    #   psycopg
    #   psycopg-pool
tzdata==2024.2 ; sys_platform == 'win32'
    # via
    #   django
    #   psycopg
whitenoise==6.8.2
    # via -r requirements.in
